requirements.txt
streamlit
pillow
openai
firebase-admin
reportlab
python-dotenv
pandas
google-cloud-vision
